+++ include/dinamic_array.hpp
@@ -15,26 +15,26 @@ class DynamicArray : public ICollection<T> {
   DynamicArray() : size_(0), data_(nullptr) {
   }
 
-  explicit DynamicArray(int size) : size_(0), data_(nullptr) {
+  explicit DynamicArray(size_t size) : size_(0), data_(nullptr) {
     if (size < 0) {
       throw InvalidArgument("Ошибка, отрицатльный размер");
     }
     Allocate(size);
   }
 
-  DynamicArray(const T* items, int count) : size_(0), data_(nullptr) {
+  DynamicArray(const T* items, size_t count) : size_(0), data_(nullptr) {
     if (count < 0) {
       throw InvalidArgument("Ошибка, отрицатльный размер");
     }
     Allocate(count);
-    for (int i = 0; i < count; ++i) {
+    for (size_t i = 0; i < count; ++i) {
       data_[i] = items[i];
     }
   }
 
   DynamicArray(const DynamicArray<T>& other) : size_(0), data_(nullptr) {
     Allocate(other.size_);
-    for (int i = 0; i < size_; ++i) {
+    for (size_t i = 0; i < size_; ++i) {
       data_[i] = other.data_[i];
     }
   }
@@ -48,19 +48,19 @@ class DynamicArray : public ICollection<T> {
     return *this;
   }
 
-  T Get(int index) const {
+  T Get(size_t index) const {
     ValidateIndex(index);
     return data_[index];
   }
 
   T Get(std::size_t index) const override {
-    if (index > static_cast<std::size_t>(static_cast<int>(index))) {
+    if (index > static_cast<std::size_t>(static_cast<size_t>(index))) {
       throw IndexOutOfRange("Ошибка, слишком большой индекс");
     }
-    return Get(static_cast<int>(index));
+    return Get(static_cast<size_t>(index));
   }
 
-  int GetSize() const {
+  size_t GetSize() const {
     return size_;
   }
 
@@ -68,50 +68,50 @@ class DynamicArray : public ICollection<T> {
     return static_cast<std::size_t>(size_);
   }
 
-  void Set(int index, const T& value) {
+  void Set(size_t index, const T& value) {
     ValidateIndex(index);
     data_[index] = value;
   }
 
-  void Resize(int new_size) {
+  void Resize(size_t new_size) {
     if (new_size < 0) {
       throw InvalidArgument("Ошибка, отрицательный размер");
     }
     std::unique_ptr<T[]> new_data = (new_size == 0 ? nullptr : std::make_unique<T[]>(new_size));
-    const int copy_count = std::min(size_, new_size);
-    for (int i = 0; i < copy_count; ++i) {
+    const size_t copy_count = std::min(size_, new_size);
+    for (size_t i = 0; i < copy_count; ++i) {
       new_data[i] = data_[i];
     }
-    for (int i = copy_count; i < new_size; ++i) {
+    for (size_t i = copy_count; i < new_size; ++i) {
       new_data[i] = T();
     }
     data_ = std::move(new_data);
     size_ = new_size;
   }
 
-  const T& operator[](int index) const {
+  const T& operator[](size_t index) const {
     ValidateIndex(index);
     return data_[index];
   }
 
-  T& operator[](int index) {
+  T& operator[](size_t index) {
     ValidateIndex(index);
     return data_[index];
   }
 
  private:
-  int size_;
+  size_t size_;
   std::unique_ptr<T[]> data_;
 
-  void Allocate(int size) {
+  void Allocate(size_t size) {
     size_ = size;
     data_ = (size == 0 ? nullptr : std::make_unique<T[]>(size));
-    for (int i = 0; i < size_; ++i) {
+    for (size_t i = 0; i < size_; ++i) {
       data_[i] = T();
     }
   }
 
-  void ValidateIndex(int index) const {
+  void ValidateIndex(size_t index) const {
     if (index < 0 || index >= size_) {
       throw IndexOutOfRange("Ошибка, индекс не в диапазоне");
     }
